added a little documentation
