+++ src/home_robot/home_robot/perception/encoders/base_encoder.py
@@ -1,3 +1,7 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
 from typing import Union
 
 from numpy import ndarray
@@ -5,6 +9,10 @@ from torch import Tensor
 
 
 class BaseImageTextEncoder:
+    """
+    Encodes images, encodes text, and allows comparisons between the two encoding.
+    """
+
     def encode_image(self, image: Union[ndarray, Tensor]):
         raise NotImplementedError
